% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{conre3}
\alias{conre3}
\title{CONRE-3 registered companies}
\format{A data frame with 133 rows and 5 variables:
\describe{
   \item{empresa}{Name of the company.}
   \item{uf}{State initials.}
   \item{cnpj}{Company national id.}
}}
\source{
\url{http://www.conre3.org.br/portal/profissionais-e-empresas-do-conre-3-2/}
}
\usage{
conre3
}
\description{
A raw dataset containing registered companies scraped from CONRE-3 site.
}
\keyword{datasets}

