% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{conre6}
\alias{conre6}
\title{CONRE-6 registered companies}
\format{An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 133 rows and 3 columns.}
\source{
\url{http://www.conre3.org.br/portal/profissionais-e-empresas-do-conre-3-2/}
}
\usage{
conre6
}
\description{
CONRE-6 registered companies
}
\keyword{datasets}

