% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{conre4}
\alias{conre4}
\title{CONRE-4 registered companies}
\format{A data frame with 42 rows and 3 variables:
\describe{
   \item{empresa}{Name of the company.}
   \item{uf}{State initials (not real).}
   \item{cnpj}{Company national id.}
}}
\source{
\url{http://www.conre4.org.br}
}
\usage{
conre4
}
\description{
CONRE-4 registered companies
}
\keyword{datasets}

