% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{conre6}
\alias{conre6}
\title{CONRE-6 registered companies}
\format{A data frame with 30 rows and 3 variables:
\describe{
   \item{empresa}{Name of the company.}
   \item{uf}{State initials.}
   \item{cnpj}{Company national id.}
}}
\source{
\url{http://www.conre3.org.br/portal/profissionais-e-empresas-do-conre-3-2/}
}
\usage{
conre6
}
\description{
CONRE-6 registered companies
}
\keyword{datasets}

