% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conre.R
\name{conre_get_cnpj}
\alias{conre_get_cnpj}
\title{Download CONRE registered companies}
\usage{
conre_get_cnpj(region = 3)
}
\arguments{
\item{region}{CONRE region from 1 to 6.}
}
\value{
\code{tibble} with 2 columns (name, state and id) as specified in the saved
  datasets.
}
\description{
Downloads CONRE registered companies from CONRE sites.
}
\seealso{
\code{\link{conre3}}, \code{\link{conre6}}
}

